more changes

+++ src/Runtime/ExecutionSession.cpp
@@ -56,7 +56,7 @@ ExecutionSession::run(
 
   std::vector<std::unique_ptr<OMTensor, decltype(&omTensorDestroy)>> outs;
 
-  for (size_t i = 0; i < omTensorListGetSize(wrappedOutput); i++) {
+  for (size_t i = 0; i < (size_t)omTensorListGetSize(wrappedOutput); i++) {
     outs.emplace_back(std::unique_ptr<OMTensor, decltype(&omTensorDestroy)>(
         omTensorListGetOmtByIndex(wrappedOutput, i), omTensorDestroy));
   }
